refactor(Emblem): 로직이 없는 Emblem을 상수 패키지에서 관리하라
