+++ src/test/java/domain/CardHandTest.java
@@ -10,6 +10,8 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
+import constant.Emblem;
+
 class CardHandTest {
 
     @Nested
