+++ src/test/java/domain/DealerTest.java
@@ -8,6 +8,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
+import constant.Emblem;
+
 public class DealerTest {
 
     @Nested
