+++ src/test/java/domain/DeckTest.java
@@ -9,6 +9,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
+import constant.Emblem;
+
 public class DeckTest {
 
     @Nested
