+++ src/test/java/domain/PlayerTest.java
@@ -11,6 +11,8 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
+import constant.Emblem;
+
 public class PlayerTest {
 
     @Nested
